+++ src/main/java/pl/peterwolf/echoesinink/progression/WorkshopRewardAllocator.java
@@ -67,8 +67,8 @@ public final class WorkshopRewardAllocator {
 				case PRESS_FRAME, MATRIX_BENCH -> rewards(RewardKind.WOODEN_MATRIX, 1);
 				case MACHINE_REMAINS, FLOOR_CACHE -> rewards(RewardKind.VILLAGE_CHRONICLE_MATRIX, 1);
 				case CELLAR_CACHE -> rewards(RewardKind.WOODEN_MATRIX, 1, RewardKind.PRINTERS_NOTES, 1);
-				case ARCHIVE_DESK -> rewards(RewardKind.DAMAGED_PAGE, small, RewardKind.BLANK_PAGE, 2 + small);
-				case INK_STATION -> rewards(RewardKind.INK_BALL, 4 + small);
+				case ARCHIVE_DESK -> rewards(RewardKind.DAMAGED_PAGE, small, RewardKind.BLANK_PAGE, 8 + small);
+				case INK_STATION -> rewards(RewardKind.INK_BALL, 10 + small);
 				case PLAQUE_CLUE -> rewards(RewardKind.MAP_FRAGMENT, 1);
 				case SUSPICIOUS_FLOOR -> rewards(RewardKind.CHARCOAL_RUBBING_PAPER, small);
 			};
@@ -76,8 +76,8 @@ public final class WorkshopRewardAllocator {
 				case PRESS_FRAME, MACHINE_REMAINS -> rewards(RewardKind.IRON_CHASE, 1);
 				case CELLAR_CACHE -> rewards(RewardKind.MISSING_HEADLINE_TYPE, 1, RewardKind.PRINTERS_NOTES, 1);
 				case FLOOR_CACHE, MATRIX_BENCH -> rewards(RewardKind.LEAD_TYPE_SET, 1);
-				case ARCHIVE_DESK -> rewards(RewardKind.PRINTERS_NOTES, 1, RewardKind.BLANK_PAGE, 2 + small);
-				case INK_STATION -> rewards(RewardKind.INK_PAD, 4 + small);
+				case ARCHIVE_DESK -> rewards(RewardKind.PRINTERS_NOTES, 1, RewardKind.BLANK_PAGE, 8 + small);
+				case INK_STATION -> rewards(RewardKind.INK_PAD, 10 + small);
 				case PLAQUE_CLUE -> rewards(RewardKind.MAP_FRAGMENT, 1);
 				case SUSPICIOUS_FLOOR -> rewards(RewardKind.METAL_TYPE, small);
 			};
@@ -86,8 +86,8 @@ public final class WorkshopRewardAllocator {
 				case MACHINE_REMAINS -> rewards(RewardKind.WOODEN_MATRIX, 1);
 				case CELLAR_CACHE -> rewards(RewardKind.PRINTERS_NOTES, 1, RewardKind.DAMAGED_PAGE, 2);
 				case FLOOR_CACHE -> rewards(RewardKind.WOODEN_MATRIX, 1);
-				case ARCHIVE_DESK -> rewards(RewardKind.DAMAGED_PAGE, 2 + small, RewardKind.BLANK_PAGE, 3 + small);
-				case INK_STATION -> rewards(RewardKind.INK_BALL, 4 + small);
+				case ARCHIVE_DESK -> rewards(RewardKind.DAMAGED_PAGE, 2 + small, RewardKind.BLANK_PAGE, 10 + small);
+				case INK_STATION -> rewards(RewardKind.INK_BALL, 10 + small);
 				case PLAQUE_CLUE -> rewards(RewardKind.MAP_FRAGMENT, 1, RewardKind.PRINTERS_NOTES, 1);
 				case SUSPICIOUS_FLOOR -> rewards(RewardKind.DAMAGED_PAGE, 1);
 			};
@@ -96,8 +96,8 @@ public final class WorkshopRewardAllocator {
 				case MACHINE_REMAINS -> rewards(RewardKind.LEAD_TYPE_SET, 1);
 				case CELLAR_CACHE -> rewards(RewardKind.IRON_CHASE, 1, RewardKind.PRINTERS_NOTES, 1);
 				case MATRIX_BENCH -> rewards(RewardKind.WOODEN_MATRIX, 1);
-				case ARCHIVE_DESK -> rewards(RewardKind.DAMAGED_PAGE, small, RewardKind.PRINTERS_NOTES, 1);
-				case INK_STATION -> rewards(RewardKind.INK_PAD, 4 + small);
+				case ARCHIVE_DESK -> rewards(RewardKind.DAMAGED_PAGE, small, RewardKind.PRINTERS_NOTES, 1, RewardKind.BLANK_PAGE, 8 + small);
+				case INK_STATION -> rewards(RewardKind.INK_PAD, 10 + small);
 				case PLAQUE_CLUE -> rewards(RewardKind.MAP_FRAGMENT, 1);
 				case SUSPICIOUS_FLOOR -> rewards(RewardKind.CHARCOAL_RUBBING_PAPER, 1 + small);
 			};